reducing code reusage using utils
