+++ app/overrides/main.html
@@ -53,48 +53,7 @@
   }
 </script>
 <script>
-  
 
-  // new DataTable('#table1', {
-  //   order: [[5, 'desc']],
-  //   "columnDefs": [
-  //       {"className": "dt-center", "targets": "_all"},
-  //       { width: '30%', targets: 0 }
-  //     ],
-  //   pageLength: 10,
-  //   layout: {
-  //       topStart: {
-  //           buttons: ['copy', 'csv', 'excel', 'pdf', 'print']
-  //       }
-  //   }
-  // });
-
-  // new DataTable('#table2', {
-  //   order: [[3, 'desc']],
-  //   "columnDefs": [
-  //       {"className": "dt-center", "targets": "_all"},
-  //       { width: '30%', targets: 0 }
-  //     ],
-  //   pageLength: 10,
-  //   layout: {
-  //       topStart: {
-  //           buttons: ['copy', 'csv', 'excel', 'pdf', 'print']
-  //       }
-  //   }
-  // });
-  // new DataTable('#table3', {
-  //   initComplete: function () {
-  //       console.log("DEBUG: DataTable initialized with", this.api().columns().count(), "columns");
-  //   },
-  //   columnDefs: [
-  //       {"className": "dt-center", "targets": "_all"},
-  //       { width: '5%', targets: 0 }
-  //   ],
-  //   paging: true,
-  //   scrollX: true,
-  //   dom: '<"export-container"Blf>rtip',
-  //   buttons: ['copy', 'csv', 'excel', 'pdf', 'print']
-  // });
   new DataTable('#table3', {
     initComplete: function () {
         this.api()
@@ -139,50 +98,6 @@
         }
     }
   });
-//   new DataTable('#table4', {
-//     initComplete: function () {
-//         this.api()
-//             .columns()
-//             .every(function () {
-//                 let column = this;
- 
-//                 // Create select element
-//                 let select = document.createElement('select');
-//                 select.add(new Option(''));
-//                 column.footer().replaceChildren(select);
- 
-//                 // Apply listener for user change in value
-//                 select.addEventListener('change', function () {
-//                     column
-//                         .search(select.value, {exact: true})
-//                         .draw();
-//                 });
-
-//                 // keep the width of the select element same as the column
-//                 select.style.width = '100%';
- 
-//                 // Add list of options
-//                 column
-//                     .data()
-//                     .unique()
-//                     .sort()
-//                     .each(function (d, j) {
-//                         select.add(new Option(d));
-//                     });
-//             });
-//     },
-//     // order: [[3, 'desc']],
-//     "columnDefs": [
-//         {"className": "dt-center", "targets": "_all"},
-//         { width: '30%', targets: 0 }
-//       ],
-//     pageLength: 10,
-//     layout: {
-//         topStart: {
-//             buttons: ['copy', 'csv', 'excel', 'pdf', 'print']
-//         }
-//     }
-//   });
 </script>
 
 {% endblock %}
