+++ app/templates/algo.html
