+++ docs/algorithms/index.md
@@ -110,7 +110,7 @@
   </style>
 </head>
 <body class="container">
-  <p><i class="footer">This page was last updated on 2025-02-27 13:50:36 UTC</i></p>
+  <p><i class="footer">This page was last updated on 2025-03-03 09:31:37 UTC</i></p>
 
   <!-- Intro Button -->
   <div class="note info" onclick="startIntro()">
