+++ docs/data/index.md
@@ -112,7 +112,7 @@
   </style>
 </head>
 <body class="container">
-  <p><i class="footer">This page was last updated on 2025-02-27 23:22:37 UTC</i></p>
+  <p><i class="footer">This page was last updated on 2025-03-03 09:31:33 UTC</i></p>
 
   <!-- Intro Button -->
   <div class="note info" onclick="startIntro()">
