+++ tests/test_sys_bio_model.py
@@ -1,38 +1,26 @@
 """
-A test markdown class for pytest unit testing.
+A test class for pytest unit testing.
 """
 
 import os
 import unittest.mock
 import pytest
 import basico
-import pandas as pd
 from vpeleaderboard.data.src.basico_model import BasicoModel
-from vpeleaderboard.data.src.sys_bio_model import SysBioModel
+from tests.utils.utils import (
+    MockModel,
+    model_fixture,
+    dynamic_model_factory,
+    temp_folder_fixture,
+    valid_sbml_folder
+)
 
-@pytest.fixture(name="model_instance")
-def model_fixture():
-    """
-    Fixture for the BasicoModel class.
-    """
-    return BasicoModel(sbml_folder_path="vpeleaderboard/data/models")
-
-@pytest.fixture(name="temp_folder")
-def temp_folder_fixture():
-    """Checks for an empty temporary directory."""
-    return "vpeleaderboard/data/models"
-
-@pytest.fixture(name="dynamic_model_creator")
-def dynamic_model_factory():
-    """
-    A fixture to create a BasicoModel instance
-    with a function-defined SBML path.
-    Each test can pass its own path as an argument.
-    """
-    def _create_model(sbml_path):
-        return BasicoModel(sbml_folder_path=sbml_path)
-
-    return _create_model
+def test_with_fixtures():
+    """Test to ensure all fixtures are accessible and not None."""
+    assert model_fixture is not None
+    assert dynamic_model_factory is not None
+    assert temp_folder_fixture is not None
+    assert valid_sbml_folder is not None
 
 def test_validate_sbml_file_path_success(tmp_path):
     """
@@ -90,23 +78,6 @@ def test_get_model_metadata():
     assert metadata["Number of Species"] >= 0
     assert metadata["Number of Parameters"] == len(basico.get_parameters())
     assert metadata["Description"] is not None
-
-class MockModel(SysBioModel):
-    """
-    Mock implementation of SysBioModel for testing.
-    """
-    def get_model_metadata(self) -> pd.DataFrame:
-        """Mock implementation returning empty DataFrame"""
-        return pd.DataFrame()
-
-@pytest.fixture(name="valid_sbml_folder_path")
-def valid_sbml_folder(tmp_path):
-    """Creates a temporary directory with a dummy XML file."""
-    folder = tmp_path / "models"
-    os.makedirs(folder, exist_ok=True)
-    with open(folder / "dummy_model.xml", "w", encoding="utf-8") as f:
-        f.write("<sbml></sbml>")
-    return folder
 
 def test_valid_sbml_directory(valid_sbml_folder_path):
     """
