+++ tests/utils/__init__.py
@@ -1,0 +1,7 @@
+"""
+Init file for the markdown_utils package.
+"""
+
+from . import utils
+
+__all__ = ["utils"]
